Deprecated types other than 64-bit signed integer for autoincrement PKs
SQLite documentation clearly states that an auto-incrementable primary key column must be a 64-bit signed integer type.

+++ examples/chrono_binding.cpp
@@ -138,7 +138,7 @@ using std::cout;
 using std::endl;
 
 struct Person {
-    int id;
+    int64 id;
     std::string name;
     std::chrono::sys_days birthdate;
 };
